refactor code: organize the directory
