+++ extension/secureflow/src/extension.ts
@@ -2,10 +2,10 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 import { SecurityIssue } from './models/security-issue';
-import { performSecurityAnalysis, performSecurityAnalysisAsync } from './security-analyzer';
-import { registerSecureFlowReviewCommand } from './git-changes';
-import { SettingsManager } from './settings-manager';
-import { WorkspaceProfilerCommand } from './workspace-profiler-command';
+import { performSecurityAnalysisAsync } from './analysis/security-analyzer';
+import { registerSecureFlowReviewCommand } from './git/git-changes';
+import { SettingsManager } from './settings/settings-manager';
+import { WorkspaceProfilerCommand } from './profiler/workspace-profiler-command';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -89,13 +89,11 @@ export function activate(context: vscode.ExtensionContext) {
 				} else {
 					// Fallback to pattern-based analysis if no API key
 					outputChannel.appendLine('⚠️ No API key found for the selected AI Model. Using pattern-based analysis only.');
-					securityIssues = performSecurityAnalysis(selectedText, aiModel);
 				}
 			} catch (error) {
 				// If there's an error with the API key or AI analysis, fallback to pattern-based
 				console.error('Error with AI analysis:', error);
 				outputChannel.appendLine(`⚠️ Error connecting to ${aiModel}: ${error}. Using pattern-based analysis only.`);
-				securityIssues = performSecurityAnalysis(selectedText, aiModel);
 			}
 			
 			// Complete the progress
